==> 1 - Data Preparation.ipynb <==
import pandas as pd
import numpy as np
import os
import glob
from pathlib import Path
import json
import re

TARGET_PATH = "./data/negative/negative_data.csv"
DATA_PATH = "./data/negative/wiki_zh"

class SentenceTokenizer():
    pattern = re.compile('([﹒﹔﹖﹗．。！？]["’”」』]{0,2}|(?=["‘“「『]{1,2}|$))')
    @staticmethod
    def sent_tokenize(article):
        sent_list = []
        for ele in SentenceTokenizer.pattern.split(article):
            if SentenceTokenizer.pattern.match(ele) and sent_list:
                sent_list[-1] += ele
            elif ele:
                sent_list.append(ele)
        return sent_list
    
def make_sentence():
    files = []
    for path in Path(DATA_PATH).rglob('wiki*'):
        if not os.path.isdir(path):
            files.append(str(path))
    lines = []
    for file in files:
        with open(file,'r', encoding = 'utf-8') as f:  
            for line in f:
                content = json.loads(line)['text']
                for sent in SentenceTokenizer.sent_tokenize(content):
                    lines.append(sent)
    pd.DataFrame(lines, columns = ['sentence']).to_csv(TARGET_PATH, index = False)

make_sentence()

positive = pd.read_excel("data/positive/negative.xlsx").iloc[:,0:1]
positive.to_csv("data/positive/positive.csv", index = False)
