==> 3 - Train Simple Model.ipynb <==
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.model_selection import train_test_split
import pandas as pd
pd.set_option('max_row', 1000)

data = pd.read_csv("data/train_data.csv")

X = data['sentence']
Y = data['label']
null_filter = X.notnull() & Y.notnull()
X = X.loc[null_filter]
Y = Y.loc[null_filter]

X_train, X_test, y_train, y_test = train_test_split(X,Y, test_size = 0.2, stratify = Y, random_state = 42)

X_TEST_TEXT = X_test.copy()

vectorizer = TfidfVectorizer( ngram_range=(1, 5),max_features = 4000)

vectorizer = CountVectorizer(ngram_range=(1, 5),max_features = 4000)

X_train = vectorizer.fit_transform(X_train)
X_test = vectorizer.transform(X_test)

from sklearn.naive_bayes import MultinomialNB

model = MultinomialNB()

model.fit(X_train, y_train)
model.score(X_test, y_test)

pred = (model.predict_proba(X_test) > 0.3)[:,1]
# pred = model.predict(X_test)

from sklearn.metrics import roc_auc_score, confusion_matrix
roc_auc_score(y_test, pred)

confusion_matrix(y_test, pred)

false_prediction = pred !=  y_test
find_sensitive = (pred == 1) & (y_test == 1)

### 没有找到的敏感词

pd.DataFrame(zip(X_TEST_TEXT[false_prediction],y_test[false_prediction], model.predict(X_test)[false_prediction]), columns = ['Text',"True Value","Prediction"])

## 找到的敏感词

pd.DataFrame(zip(X_TEST_TEXT[find_sensitive],y_test[find_sensitive], pred[find_sensitive]), columns = ['Text',"True Value","Prediction"])

from sklearn.pipeline import Pipeline

X_train, X_test, y_train, y_test = train_test_split(X,Y, test_size = 0.2, stratify = Y, random_state = 42)
X_TEST_TEXT = X_test.copy()

pipe = Pipeline([('vec',TfidfVectorizer( ngram_range=(1, 5),max_features = 4000)),
         ('model', MultinomialNB(class_prior=[0.2,0.1]))])


pipe = Pipeline([('vec',CountVectorizer( ngram_range=(1, 5),max_features = 4000)),
         ('model', MultinomialNB(class_prior=[0.2,0.1]))])

pipe.fit(X_train, y_train)

roc_auc_score(y_test, pipe.predict(X_test))

pipe.predict(['那个 气枪 我 想要'])

pipe.predict(['你 特么 太 淫荡 了'])
